requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-text-splitters
langchain-groq
groq
faiss-cpu
sentence-transformers
pypdf